Refactor test classes and package naming
Aim is to express previously implemented end-to-end test upgrade.
