+++ src/test/java/com/ibardos/motoShop/MotoShopApplicationGeneralTests.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationContext;
 
 @SpringBootTest
-class MotoShopApplicationTests {
+class MotoShopApplicationGeneralTests {
     @Autowired
     ApplicationContext applicationContext;
