+++ src/test/java/com/ibardos/motoShop/endToendTests/apiTests/MotorcycleStockControllerApiTests.java
@@ -1,4 +1,4 @@
-package com.ibardos.motoShop.controller;
+package com.ibardos.motoShop.endToendTests.apiTests;
 
 import org.junit.jupiter.api.*;
 
@@ -20,14 +20,14 @@ import java.nio.file.Path;
 import java.nio.file.Files;
 
 /**
- * End-to-end tests against API endpoints in ManufacturerController.
+ * End-to-end tests against API endpoints in MotorcycleStockController.
  */
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.ANY)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql(scripts = "/schema.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 @Sql(scripts = "/data.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class ManufacturerControllerTest {
+public class MotorcycleStockControllerApiTests {
     @LocalServerPort
     private int port;
 
@@ -40,16 +40,16 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(1)
-    void add_newValidManufacturer_statusCode201WithProperJson() throws Exception {
+    void add_newValidMotorcycleStock_statusCode201WithProperJson() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/add";
+        String url = baseUrl + port + "/motorcycle/stock/add";
 
         int expectedResponseStatus = 201;
-        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/responses/Add.json")));
+        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/responses/Add.json")));
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
-                .POST(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/requests/AddValid.json")))
+                .POST(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/requests/AddValid.json")))
                 .build();
 
         // Act
@@ -65,15 +65,15 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(2)
-    void add_newInvalidManufacturer_statusCode400() throws Exception {
+    void add_newInvalidMotorcycleStock_statusCode400() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/add";
+        String url = baseUrl + port + "/motorcycle/stock/add";
 
         int expectedResponseStatus = 400;
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
-                .POST(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/requests/AddInvalid.json")))
+                .POST(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/requests/AddInvalid.json")))
                 .build();
 
         // Act
@@ -87,12 +87,12 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(3)
-    void get_manufacturerWithValidId_statusCode200WithProperJson() throws Exception {
+    void get_motorcycleStockWithValidId_statusCode200WithProperJson() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/get/6";
+        String url = baseUrl + port + "/motorcycle/stock/get/8";
 
         int expectedResponseStatus = 200;
-        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/responses/Get.json")));
+        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/responses/Get.json")));
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .GET()
@@ -111,9 +111,9 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(4)
-    void get_manufacturerWithInvalidId_statusCode404() throws Exception {
+    void get_motorcycleStockWithInvalidId_statusCode404() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/get/55";
+        String url = baseUrl + port + "/motorcycle/stock/get/55";
 
         int expectedResponseStatus = 404;
 
@@ -132,12 +132,12 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(5)
-    void getAll_listOfManufacturers_statusCode200WithProperJson() throws Exception {
+    void getAll_listOfMotorcycleStocks_statusCode200WithProperJson() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/get/all";
+        String url = baseUrl + port + "/motorcycle/stock/get/all";
 
         int expectedResponseStatus = 200;
-        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/responses/GetAll.json")));
+        String expectedResponseBody = new String(Files.readAllBytes(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/responses/GetAll.json")));
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .GET()
@@ -156,9 +156,9 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(6)
-    void getAll_listOfManufacturersFromInvalidUrl_statusCode400() throws Exception {
+    void getAll_listOfMotorcycleStocksFromInvalidUrl_statusCode400() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/get/allInvalid";
+        String url = baseUrl + port + "/motorcycle/stock/get/allInvalid";
 
         int expectedResponseStatus = 400;
 
@@ -177,15 +177,15 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(7)
-    void update_manufacturerWithValidId_statusCode204() throws Exception {
+    void update_motorcycleStockWithValidId_statusCode204() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/update";
+        String url = baseUrl + port + "/motorcycle/stock/update";
 
         int expectedResponseStatus = 204;
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
-                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/requests/UpdateValid.json")))
+                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/requests/UpdateValid.json")))
                 .build();
 
         // Act
@@ -199,15 +199,15 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(8)
-    void update_manufacturerWithInvalidId_statusCode404() throws Exception {
+    void update_motorcycleStockWithInvalidId_statusCode404() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/update";
+        String url = baseUrl + port + "/motorcycle/stock/update";
 
         int expectedResponseStatus = 404;
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
-                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/requests/UpdateInvalidId.json")))
+                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/requests/UpdateInvalidId.json")))
                 .build();
 
         // Act
@@ -221,15 +221,15 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(9)
-    void update_manufacturerWithInvalidJson_statusCode400() throws Exception {
+    void update_motorcycleStockWithInvalidJson_statusCode400() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/update";
+        String url = baseUrl + port + "/motorcycle/stock/update";
 
         int expectedResponseStatus = 400;
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
-                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/manufacturer/requests/UpdateInvalidJson.json")))
+                .PUT(HttpRequest.BodyPublishers.ofFile(Path.of("src/test/resources/jsonsForUnitTests/motorcycleStock/requests/UpdateInvalidJson.json")))
                 .build();
 
         // Act
@@ -243,9 +243,9 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(10)
-    void delete_manufacturerWithValidId_statusCode204() throws Exception {
+    void delete_motorcycleStockWithValidId_statusCode204() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/delete/5";
+        String url = baseUrl + port + "/motorcycle/stock/delete/9";
 
         int expectedResponseStatus = 204;
 
@@ -265,33 +265,11 @@ public class ManufacturerControllerTest {
 
     @Test
     @Order(11)
-    void delete_manufacturerWithInvalidId_statusCode404() throws Exception {
+    void delete_motorcycleStockWithInvalidId_statusCode404() throws Exception {
         // Arrange
-        String url = baseUrl + port + "/manufacturer/delete/33";
+        String url = baseUrl + port + "/motorcycle/stock/delete/99";
 
         int expectedResponseStatus = 404;
-
-        HttpRequest request = HttpRequest.newBuilder(URI.create(url))
-                .headers("Content-Type", "application/json")
-                .DELETE()
-                .build();
-
-        // Act
-        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
-
-        int resultResponseStatus = response.statusCode();
-
-        // Assert
-        assertEquals(expectedResponseStatus, resultResponseStatus);
-    }
-
-    @Test
-    @Order(12)
-    void delete_manufacturerWithIdHasForeignKeyRestriction_statusCode500() throws Exception {
-        // Arrange
-        String url = baseUrl + port + "/manufacturer/delete/1";
-
-        int expectedResponseStatus = 500;
 
         HttpRequest request = HttpRequest.newBuilder(URI.create(url))
                 .headers("Content-Type", "application/json")
